Move test_movetree from test_payu to test_fsops.py
